Procedures for Content update and Final Animation
This document contains procedures on how to update and add content to
the HeroesforStudents.org website via GoDaddy account.  It includes
screenshots and arrows pointing to key areas of focus needed in order to
perform steps in the various procedures.

This also marks the final upload of the animation video onto Vimeo which
took place on Friday, April 25, 2015.